refactor: 매수,매도 데이터 타입 변경

+++ src/main/java/dev/syntax/domain/investment/service/TradeOrderService.java
@@ -5,10 +5,10 @@ import dev.syntax.domain.investment.entity.TradeOrder;
 public interface TradeOrderService {
     public TradeOrder buy(
             String cano,          // CHAR(8)
-            Long userId,          // BIGINT
+            Long userId,
             String productCode,   // pdno
             String productName,   // prdt_name
-            int quantity,         // ord_qty (INT)
+            long quantity,         // ord_qty (INT)
             long price            // ord_unpr (BIGINT)
     );
     public TradeOrder sell(
@@ -16,7 +16,7 @@ public interface TradeOrderService {
             Long userId,
             String productCode,
             String productName,
-            int quantity,
+            long quantity,
             long price
     );
 }
